auction controller updated

+++ controllers/Auctioncontroller.js
@@ -6,12 +6,29 @@ const AppError = require('./../utils/appError');
 
 exports.createAuction = catchAsync(async (req, res, next) => {
   const { timeLine } = req.body;
+  console.log('timeLine :>> ', timeLine);
 
-  //* create a expirey clime date exactly 30-days after the timeLine
+  const timeline = new Date();
+  timeline.setHours(new Date().getHours() + 24 * timeLine);
+
+  // console.log('timeLine-Date:>> ', timeline.getDate());
+  // console.log('timeLine-Day:>> ', timeline.getDay());
+  // console.log('timeLine-Month:>> ', timeline.getMonth());
+  // console.log('timeLine-Year:>> ', timeline.getFullYear());
+
+  const claimExpiry = new Date(timeline);
+  claimExpiry.setHours(new Date().getHours() + 24 * 30);
+
+  // console.log('claimExpiry-Date:>> ', claimExpiry.getDate());
+  // console.log('claimExpiry-Day:>> ', claimExpiry.getDay());
+  // console.log('claimExpiry-Month:>> ', claimExpiry.getMonth());
+  // console.log('claimExpiry-Year:>> ', claimExpiry.getFullYear());
 
   const auction = await Auction.create({
-    user: req.user._id,
     ...req.body,
+    user: req.user._id,
+    claimExpiry: claimExpiry,
+    timeLine: timeline,
   });
 
   res.status(200).json({
@@ -21,8 +38,17 @@ exports.createAuction = catchAsync(async (req, res, next) => {
 });
 
 exports.getAllAuctions = catchAsync(async (req, res, next) => {
-  //* globally only get the published-auctions
-  const auctions = await Auction.find({ status: 'published' });
+  let auctions = await Auction.find({ status: 'published' });
+
+  res.status(200).json({
+    status: 'success',
+    results: auctions.length,
+    auctions,
+  });
+});
+
+exports.myAuctions = catchAsync(async (req, res, next) => {
+  const auctions = await Auction.find({ user: req.user._id });
 
   res.status(200).json({
     status: 'success',
@@ -41,6 +67,23 @@ exports.getAuction = catchAsync(async (req, res, next) => {
         404
       )
     );
+
+  // console.log('timeLine-Date:>> ', auction.timeLine.getDate());
+  // console.log('timeLine-Day:>> ', auction.timeLine.getDay());
+  // console.log('timeLine-Month:>> ', auction.timeLine.getMonth());
+  // console.log('timeLine-Year:>> ', auction.timeLine.getFullYear());
+
+  // console.log('claimExpiry-Date:>> ', auction.claimExpiry.getDate());
+  // console.log('claimExpiry-Day:>> ', auction.claimExpiry.getDay());
+  // console.log(
+  //   'claimExpiry-Month:>> ',
+  //   auction.claimExpiry.getMonth()
+  // );
+  // console.log(
+  //   'claimExpiry-Year:>> ',
+  //   auction.claimExpiry.getFullYear()
+  // );
+
   res.status(200).json({
     status: 'success',
     auction,
@@ -49,7 +92,10 @@ exports.getAuction = catchAsync(async (req, res, next) => {
 
 //* publish-Auction
 exports.publishAuction = catchAsync(async (req, res, next) => {
-  const auction = await Auction.findById(req.params.id);
+  const auction = await Auction.findById({
+    user: req.user._id,
+    _id: req.params.id,
+  });
 
   if (!auction)
     return next(
@@ -92,8 +138,24 @@ exports.claimAuction = catchAsync(async (req, res, next) => {
 });
 
 exports.updateAuction = catchAsync(async (req, res, next) => {
+  const auction = await Auction.findById(req.params.id);
+
+  if (!auction) {
+    return next(
+      new AppError(
+        `Can't find any auction with id ${req.params.id}`,
+        404
+      )
+    );
+  }
+  if (auction.status !== 'inProgress') {
+    return next(
+      new AppError(`you can Only update the inProgress auctions`, 400)
+    );
+  }
+
   const updateAuction = await Auction.findByIdAndUpdate(
-    { user: req.user._id, status: 'inProgress' },
+    { _id: req.params.id, status: 'inProgress' },
     { ...req.body },
     {
       runValidators: true,
@@ -104,7 +166,7 @@ exports.updateAuction = catchAsync(async (req, res, next) => {
   if (!updateAuction)
     return next(
       new AppError(
-        `Can't find any auction with id ${req.user._id}`,
+        `Can't find any auction with id ${req.params.id}`,
         404
       )
     );
@@ -116,12 +178,28 @@ exports.updateAuction = catchAsync(async (req, res, next) => {
 });
 
 exports.deleteAuction = catchAsync(async (req, res, next) => {
-  const deletedUser = await Auction.findByIdAndDelete({
-    user: req.user._id,
+  const auction = await Auction.findById(req.params.id);
+
+  if (!auction) {
+    return next(
+      new AppError(
+        `Can't find any auction with id ${req.params.id}`,
+        404
+      )
+    );
+  }
+  if (auction.status !== 'inProgress') {
+    return next(
+      new AppError(`you can Only delete the inProgress auctions`, 400)
+    );
+  }
+
+  const deleteAuction = await Auction.findByIdAndDelete({
+    _id: req.params.id,
     status: 'inProgress',
   });
 
-  if (!deletedUser)
+  if (!deleteAuction)
     return next(
       new AppError(
         `No Auction found against id ${req.params.id}`,
@@ -131,7 +209,33 @@ exports.deleteAuction = catchAsync(async (req, res, next) => {
 
   res.status(200).json({
     status: 'success',
-    user: deletedUser,
+    auction: deleteAuction,
+  });
+});
+
+//* BID
+exports.createBid = catchAsync(async (req, res, next) => {
+  const { biddingPrice } = req.body;
+  const { id } = req.params;
+
+  const auction = await Auction.findById(id);
+  if (!auction) {
+    return next(
+      new AppError(`Can't find any auction with id ${id}`, 404)
+    );
+  }
+
+  const bid = await Bid.create({
+    user: req.user._id,
+    biddingPrice,
+  });
+
+  auction.bids.unshift(bid._id);
+  await auction.save();
+
+  res.status(200).json({
+    status: 'success',
+    bid,
   });
 });
